refactor: Move OntologyCacheHelpers and OntologyTriplestoreHelpers into CoreModule
